Add files via upload
The aim of the Plane and Missile Game Project is to create a simple yet engaging digital game that challenges players’ reflexes and decision-making skills. The game is designed for a broad audience, including children, casual gamers, and developers learning basic game programming. It incorporates elements of strategy and reaction to provide a fun, interactive, and educational experience.